frontend/requirements.txt
PySide6>=6.6.0
soundcard>=0.4.2
webrtcvad>=2.0.10
openai>=1.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
requests>=2.32.0
